Re-organized classes and trigger warnings for deprecated ones (task #5452)
